STT_Alibaba/pyproject.toml
[project]
name = "stt-alibaba"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.13",
    "google>=3.0.0",
    "google-genai>=1.66.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
]
